More envconfig tweaks
